requirements.txt
requests
pandas
numpy
scikit-learn
statsmodels
xgboost
matplotlib
seaborn
pyyaml